Moving out account settings selectors and removing no longer used hoc
